Add the project images in the assets and start to create the first flow.
Add the GitHub images in the assets folder.

Create the MainCoordinator to provide navigation to other coordinators. The Main Coordinator is responsible for manage the main app flow, which means that this component will be responsible to start and present all the main flow branches'.

The LoginViewController was created with the LoginView. The LoginView is the view that contains all the login layout. This approach is the same presented by Guilherme Rambo in his article: 'MVC - Many View Controllers'.