Change GetFollowersView documentation, create NavigationCoordinator, create ErrorMessage enumeration.
The GetFollowersView's documentation was updated. The action getFollowersButtonTapped were added and also a new mark(action).

The protocol GetFollowersViewControllerDelegate is now documented.

The GetFollowersError property, errorMessage, was renamed to message.

The marks defining the code blocks contexts were added to CustomAlertView.

The NavigationCoordinator protocol was created to provide a coordinator with a navigation controller. This navigation controller is responsible for embed the coordinator's flow.