Move the Get Followers scene to Get Followers module.
The Get Followers module was created to separate the feature Get Followers from the application's main module. That is based on modular application architecture. At first, the feature was in the main module, but the modularization was adopted for learning purposes.

OBS: The modularization adopted here is based on my current job(until date 27/03/2020). Here, I'm using the same idea, but with a different approach.