Create the Networking communication interface and URI configuration.
The Communication Interface is responsible to communicate directly to Networking Service.

The Networking Service is responsible to send a request where a task is created using the URLSession.

The URI is built through the type Request:
HTTPMethod: enumeration contains the request methods.
Endpoint: contains the data relative to the URL.
Body: request body.

NetworkingResponse and Networking error are the types returned by the service.